refactor: move customer service hooks
